Move ToString to util
