+++ android/framework/util/CMakeLists.txt
@@ -38,6 +38,13 @@ target_sources(gfxrecon_util
                    ${GFXRECON_SOURCE_DIR}/framework/util/options.cpp
                    ${GFXRECON_SOURCE_DIR}/framework/util/shared_mutex.h
                    ${GFXRECON_SOURCE_DIR}/framework/util/shared_mutex.cpp
+                   ${GFXRECON_SOURCE_DIR}/framework/util/custom_vulkan_to_string.h
+                   ${GFXRECON_SOURCE_DIR}/framework/util/custom_vulkan_to_string.cpp
+                   ${GFXRECON_SOURCE_DIR}/framework/generated/generated_vulkan_enum_to_string.h
+                   ${GFXRECON_SOURCE_DIR}/framework/generated/generated_vulkan_enum_to_string.cpp
+                   ${GFXRECON_SOURCE_DIR}/framework/generated/generated_vulkan_pnext_to_string.cpp
+                   ${GFXRECON_SOURCE_DIR}/framework/generated/generated_vulkan_struct_to_string.h
+                   ${GFXRECON_SOURCE_DIR}/framework/generated/generated_vulkan_struct_to_string.cpp
               )
 
 target_compile_definitions(gfxrecon_util
@@ -52,5 +59,6 @@ target_include_directories(gfxrecon_util
 
 target_link_libraries(gfxrecon_util
                           platform_specific
+                          vulkan_registry
                           ${GFXRECON_SOURCE_DIR}/external/precompiled/android/lib/${ANDROID_ABI}/liblz4_static.a
                           z)
